requirements.pip.run.txt
aiohttp
libvirt-python
pyyaml
paramiko
setuptools
